Split out most ws_socket code from socket code
