+++ mutiny-core/src/networking/mod.rs
@@ -2,3 +2,6 @@ pub mod socket;
 
 #[cfg(target_arch = "wasm32")]
 pub mod proxy;
+
+#[cfg(target_arch = "wasm32")]
+pub mod ws_socket;
